disconnectMessage2 {shinydisconnect}	R Documentation

Show a nice message when a shiny app disconnects or errors

Description

This function is a version of disconnectMessage() with a pre-set combination of parameters that results in a large centered message.

Usage

disconnectMessage2()

[Package shinydisconnect version 0.1.0 Index]